deze doet het weer

+++ SomerenModel/Drink.cs
@@ -12,7 +12,7 @@ namespace SomerenModel
         public string DrinkName { get; set; } // Drinkname, e.g. Fanta, Cola
         public decimal DrinkPrice { get; set; } // DrinkPrice, e.g. Math, 3.99, 5.00
         public int DrinkStock { get; set; } // DrinkStock, e.g. 20, 60
-        public decimal DrinkVAT { get; set; } // DrinkVAT, e.g. 21%
+        public int DrinkVAT { get; set; } // DrinkVAT, e.g. 21%
         public decimal DrinkValue { get; set; } //money earned from the drink probably?
         public int DrinksSold { get; set; } //amount of drinks sold
         public bool StockAmount {get; set;}
